lista dostepnych rezerwacji. work in progress 1

+++ src/main/java/semato/semato_med/repository/PhysicianRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import semato.semato_med.model.Physician;
 
 @Repository
-public interface PhysicianRepository extends JpaRepository<Physician, Integer> {
+public interface PhysicianRepository extends JpaRepository<Physician, Long> {
 }
